Move controllers integration tests next to controllers
