+++ controllers/controllers/networking/cfroute_controller_test.go
@@ -1,4 +1,4 @@
-package integration_test
+package networking_test
 
 import (
 	"context"
