+++ controllers/controllers/networking/suite_test.go
@@ -1,4 +1,4 @@
-package integration_test
+package networking_test
 
 import (
 	"context"
@@ -47,8 +47,8 @@ var _ = BeforeSuite(func() {
 	testEnv = &envtest.Environment{
 		// TODO: Reconcile with CRDInstallOptions
 		CRDDirectoryPaths: []string{
-			filepath.Join("..", "..", "..", "..", "helm", "controllers", "templates", "crds"),
-			filepath.Join("..", "..", "..", "..", "tests", "vendor", "contour"),
+			filepath.Join("..", "..", "..", "helm", "controllers", "templates", "crds"),
+			filepath.Join("..", "..", "..", "tests", "vendor", "contour"),
 		},
 		ErrorIfCRDPathMissing: true,
 	}
