+++ controllers/controllers/services/cfservicebinding_controller_test.go
@@ -1,4 +1,4 @@
-package integration_test
+package services_test
 
 import (
 	"context"
